Replace CustomResolveVisitor by NoInlineAnnotationTransformationResolveVisitor
